refactor: viewport highlight ranges
